Replace `flow-example` with flow-example-parent-children
