--- sdk-integration-examples/web-sdk/react/flow-example-parent-child/src/components/Details/Details.css
@@ -1,138 +1,0 @@
-.details {
-  width: 100%;
-  min-height: 100vh;
-  background-color: #fff;
-  z-index: initial;
-}
-
-.details.on-top {
-  z-index: 999;
-}
-
-.details-header {
-  padding: 15px 20px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  align-content: center;
-  gap: 15px;
-  box-shadow: 0 1px 4px rgba(0, 0, 0, 0.1);
-  position: sticky;
-  top: 0;
-  background: #fff;
-  z-index: 10;
-}
-
-.details-header-icon {
-  cursor: pointer;
-  width: 32px;
-  height: 32px;
-}
-
-.details-header-title {
-  font-size: 24px;
-  font-weight: 600;
-  margin: 0;
-  color: #000000;
-}
-
-.details-content {
-  padding: 20px;
-}
-
-.details-section-title {
-  font-size: 18px;
-  font-weight: 500;
-  margin: 25px 0 15px 0;
-  color: #000000;
-  text-transform: capitalize;
-}
-
-.details-form-control {
-  margin-bottom: 15px;
-  border: 1px solid #dedede;
-  border-radius: 8px;
-  padding: 8px 12px;
-  background: #fff;
-}
-
-.details-form-label {
-  text-transform: capitalize;
-  font-size: 12px;
-  font-weight: 400;
-  color: #666;
-  margin-bottom: 4px;
-  display: block;
-}
-
-.details-form-input {
-  color: #000000;
-  border: 0;
-  background: transparent;
-  width: 100%;
-  font-size: 16px;
-  font-weight: 400;
-  padding: 0;
-  height: 24px;
-  outline: none;
-}
-
-.details-form-input:disabled {
-  color: #000;
-  opacity: 1;
-}
-
-.details-consent-checkbox {
-  display: flex;
-  align-items: flex-start;
-  gap: 10px;
-  margin: 25px 0 30px 0;
-}
-
-.details-consent-checkbox input[type='checkbox'] {
-  margin-top: 3px;
-}
-
-.details-consent-text {
-  font-size: 14px;
-  color: #666;
-  margin: 0;
-  line-height: 1.4;
-}
-
-.details-submit-button {
-  width: 100%;
-  height: 48px;
-  border-radius: 8px;
-  background: #051c2c;
-  color: #fff;
-  font-size: 16px;
-  font-weight: 500;
-  border: 0;
-  cursor: pointer;
-  margin-bottom: 10px;
-}
-
-.details-submit-button:disabled {
-  background: gray;
-}
-
-.details-submit-button:not(:disabled):hover {
-  background: #0a2c44;
-}
-
-@media (min-width: 768px) {
-  .details {
-    max-width: 600px;
-    margin: 0 auto;
-  }
-
-  .details-content {
-    padding: 40px 20px;
-  }
-
-  .details-submit-button {
-    max-width: 100%;
-    margin: 0 auto;
-  }
-}
