--- sdk-integration-examples/web-sdk/react/flow-example-parent-child/src/vite-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="vite/client" />
