--- sdk-integration-examples/web-sdk/react/flow-example-parent-child/vite.config.ts
@@ -1,13 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-import basicSsl from "@vitejs/plugin-basic-ssl";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react(), basicSsl()],
-  server: {
-    fs: {
-      allow: ["../../.."],
-    },
-  },
-});
