+++ sdk-integration-examples/web-sdk/react/flow-example/.env.local.example
@@ -1,0 +1,3 @@
+VITE_SDK_SESSION_URL=https://...
+VITE_SDK_SESSION_TOKEN=idv_...
+VITE_SDK_SESSION_BUILD_ID=build_...
