+++ sdk-integration-examples/web-sdk/react/flow-example/README.md
@@ -15,3 +15,19 @@ You may also see any lint errors in the console.
 ### `npm run build`
 
 Builds the app for production to the `dist` folder.\
+
+### `env variables`
+
+Create a `.env.local` file at the root of this project
+
+With the following content (use your data in the placeholders)
+
+```bash
+VITE_SDK_SESSION_URL=https://........
+VITE_SDK_SESSION_TOKEN=idv_..........
+VITE_SDK_SESSION_BUILD_ID=build_.....
+```
+
+### `Note`
+
+Dont forget to place `idverse-sdk-browser` as per `package.json` indicates
