+++ sdk-integration-examples/web-sdk/react/flow-example/src/components/Intro.tsx
