+++ sdk-integration-examples/web-sdk/react/flow-example/src/main.tsx
@@ -1,14 +1,14 @@
-import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
-import "./index.css";
+import ReactDOM from 'react-dom/client';
+import { App } from './App.tsx';
+import './index.css';
 
-import { JSX as LocalJSX } from "@idverse/idverse-sdk-browser/ui";
-import { defineCustomElements } from "@idverse/idverse-sdk-browser/ui/loader";
-import { HTMLAttributes } from "react";
+import { JSX as LocalJSX } from '@idverse/idverse-sdk-browser/ui';
+import { defineCustomElements } from '@idverse/idverse-sdk-browser/ui/loader';
+import { HTMLAttributes } from 'react';
 
 type StencilToReact<T> = {
   [P in keyof T]?: T[P] &
-    Omit<HTMLAttributes<Element>, "className"> & {
+    Omit<HTMLAttributes<Element>, 'className'> & {
       class?: string;
     };
 };
@@ -20,7 +20,7 @@ declare global {
   }
 }
 
-ReactDOM.createRoot(document.getElementById("root")!).render(<App />);
+ReactDOM.createRoot(document.getElementById('root')!).render(<App />);
 
 // Initiate IDVerseSDK UI components.
 defineCustomElements();
